remove testingPopover, poster-competition, img carousel
