+++ server/api/iotd.ts
@@ -1,5 +1,18 @@
+import * as z from 'zod'
 import logger from '../utils/logger'
-import type { NasaImgType } from '@/types/nasa'
+
+export const NasaImgSchema = z.object({
+  title: z.string(),
+  explanation: z.string(),
+  date: z.string(),
+  url: z.string(),
+  hdurl: z.string(),
+  media_type: z.string(),
+  copyright: z.string(),
+  service_version: z.string()
+})
+
+export type NasaImgT = z.infer<typeof NasaImgSchema>
 
 export default defineEventHandler(async (event) => {
   // use date for KV storage
@@ -8,12 +21,12 @@ export default defineEventHandler(async (event) => {
   const storage = useStorage('data')
 
   // Check if item exists in storage
-  let nasaImg: NasaImgType | null = null
+  let nasaImg: NasaImgT | null = null
   const isItemInStorage = await storage.hasItem(cacheKey)
   logger.info(`Is ${cacheKey} in storage: ${isItemInStorage}`)
   if (isItemInStorage) {
     // Get item from storage
-    nasaImg = await storage.getItem<NasaImgType>(cacheKey)!
+    nasaImg = await storage.getItem<NasaImgT>(cacheKey)!
     logger.info(`Got ${cacheKey} from storage: `, nasaImg)
   } else {
     // Get the item for the previous day
@@ -22,7 +35,7 @@ export default defineEventHandler(async (event) => {
 
     const isPreviousItemInStorage = await storage.hasItem(previousCacheKey)
     if (isPreviousItemInStorage) {
-      nasaImg = await storage.getItem<NasaImgType>(previousCacheKey)!
+      nasaImg = await storage.getItem<NasaImgT>(previousCacheKey)!
       logger.info(`Got ${previousCacheKey} from storage: `, nasaImg)
     }
   }
