--- types/testing.ts
@@ -1,3 +1,0 @@
-export type TestingSettingsType = {
-  authOn: boolean
-}
